+++ plugin/Hydro/FlowCalculations.cs
@@ -12,6 +12,21 @@ using Rhino.Geometry;
 
 public static class FlowCalculations
 {
+    public static double getSensibleFidelity(List<Point3d> flowPathPoints)
+    {
+        // Measure distances between each of the points
+        double[] distances = new double[flowPathPoints.Count];
+        for (var i = 1; i < flowPathPoints.Count; i++)
+        {
+            distances[i - 1] = flowPathPoints[0].DistanceTo(flowPathPoints[i]);
+        }
+        // Sort and find the average of the first three results; then divide to find a decent result
+        Array.Sort(distances);
+        var averageDistance = (distances[0] + distances[1] + distances[2]) / 3;
+        var sensibleDistance = averageDistance / 10;
+        return sensibleDistance;
+    }
+
     public static Tuple<Grasshopper.DataTree<object>, List<Polyline>> MakeOutputs(List<Point3d>[] flowPathPoints)
     {
         var allFlowPathPointsTree = new Grasshopper.DataTree<object>();